requirements.txt
openai
ollama
swarm
autogen
python-dotenv
flask
